w3c-reports/12092014/efl/runned-tests/html/webappapis/scripting/processing-model-2/body-onerror-runtime-error-expected.txt

CONSOLE MESSAGE: line 27: ReferenceError: Can't find variable: undefined_variable

FAIL <body onerror> - runtime error in <script> assert_true: ran expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/html/webappapis/scripting/processing-model-2/body-onerror-runtime-error.html:31:20
step@http://localhost:8800/resources/testharness.js:798:30
global code@http://localhost:8800/html/webappapis/scripting/processing-model-2/body-onerror-runtime-error.html:30:11)
PASS <body onerror> - runtime error in <script> (column) 
